Move terrain to own class
